app/requirements.txt
django
djangorestframework
django-cors-headers
environs
psycopg2